backend/requirements.txt
asgiref==3.7.2
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.11.2
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.6
PyJWT==2.8.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0